refactor: Split `test_views.py` into tests per ViewSet
I am hoping this improves test discoverability, reduces the risk
of merge conflicts, and reduces cognitive load when making changes.

This is a simple copy-and-paste to different files, no test changes!
It's also not a perfect output, some files are still quite large
but they at least reflect a one-to-one relationship between test
file and viewset. Where the test files are large, they might indicate
that the viewset is a bit too all-ecompassing, e.g.,
'response-demographic-aggregations' might be better represented as
its own viewset rather than part of the ResponseViewSet.